% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{Ds_fun}
\alias{Ds_fun}
\title{Compute Jacobian of Market Shares}
\usage{
Ds_fun(x)
}
\arguments{
\item{x}{an object that has a Ds_fun method}
}
\value{
x with the Jacobian of shares attached
}
\description{
Compute Jacobian of Market Shares
}
\examples{
#Ds_fun(x)
}
\seealso{
\code{\link{Ds_fun.Logit_Demand_Market}}
}
