% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm_profits.R
\name{firm_profits}
\alias{firm_profits}
\title{Title Compute Firm Profits}
\usage{
firm_profits(x)
}
\arguments{
\item{x}{a 'Logit_Demand_Market' object}
}
\value{
x with the component 'firm_profits' added, a vector of the profits of each firm given the data in x
}
\description{
Title Compute Firm Profits
}
\examples{
#NA
}
