% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassConstructor.R
\name{ldmkt_construct}
\alias{ldmkt_construct}
\title{Logit Demand Class Constructor}
\usage{
ldmkt_construct(Firms, Prod_ids = NULL, Delta = NULL, Price = NULL,
  Share = NULL, Struct_error = NULL, Mc_fixed = NULL, Mc_error = NULL,
  Markup = NULL, Deriv_price = 1, U_out_opt = 0, Market_Size = 1)
}
\arguments{
\item{Firms}{vector giving which firm owns each product in the market}

\item{Prod_ids}{unique ids for the products in the market}

\item{Delta}{fixed part of consumer utility for each product}

\item{Price}{the prices of each product}

\item{Share}{the market share of each product}

\item{Struct_error}{the econometric error for each product}

\item{Mc_fixed}{the fixed part of marginal cost for each product}

\item{Mc_error}{the econometric error in marginal cost}

\item{Markup}{the markup for each product}

\item{Deriv_price}{the derivative of the consumers utility function with respect to price}

\item{U_out_opt}{the utility of the outside option (default = 0)}

\item{Market_Size}{the size of the market (default to 0)}
}
\value{
will return the number of products (this is a constructor, do not use!)
}
\description{
This is the constructor for an R6 class, it does not have a return that will be useable. It defualts all arguements to 0 if not given.
Objects of the Logit Demand Class carry all needed information to find counterfactual prices, and market shares, using their methods.
}
\examples{
#NA
}

