% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassMethods.R
\name{ldmkt_Ds_fun}
\alias{ldmkt_Ds_fun}
\title{Jacobian of market shares}
\usage{
ldmkt_Ds_fun()
}
\value{
invisibly return self
}
\description{
compute matrices related to the jacobian of market shares with respect to prices
}
\examples{
#NA
}

