% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassExpProf.R
\name{ldmkt_exp_profits}
\alias{ldmkt_exp_profits}
\title{Expected Variable Profits}
\usage{
ldmkt_exp_profits(mc_error_fun, struct_error_fun, draws, tol)
}
\arguments{
\item{mc_error_fun}{function for drawing from the marginal cost errors}

\item{struct_error_fun}{function for drawing from the utility errors}

\item{draws}{the number of draws to take}

\item{tol}{when should the zeta fixed point converge}
}
\value{
vector of each firms expected profits
}
\description{
Given a pair of functions for drawing from the distribution of the econometric errors,
 draw from the distribution, run the Zeta Fixed Point to re-compute prices and profits for each draw,
 and find the expected variable profits.
}
\examples{
#NA
}

