% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassMethods.R
\name{ldmkt_firm_profits}
\alias{ldmkt_firm_profits}
\title{Firm Profits}
\usage{
ldmkt_firm_profits()
}
\value{
return the firm profits
}
\description{
compute the profits of the firms, set this private attribute, and return the profits
}
\examples{
#NA
}

