% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassMethods.R
\name{ldmkt_marginal_cost}
\alias{ldmkt_marginal_cost}
\title{Marginal Cost method}
\usage{
ldmkt_marginal_cost()
}
\value{
invisible return self
}
\description{
Compute marginal costs by subtracting the markup from the price
}
\examples{
#NA
}

