% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassMethods.R
\name{ldmkt_markups}
\alias{ldmkt_markups}
\title{Markup method}
\usage{
ldmkt_markups()
}
\value{
no return
}
\description{
Compute the markups over marginal cost using the first order condition for NE prices, and current market shares.
}
\examples{
#NA
}

