% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassMethods.R
\name{ldmkt_share}
\alias{ldmkt_share}
\title{Share Method}
\usage{
ldmkt_share()
}
\value{
invisible return self
}
\description{
Computes the market shares of products using the sum of delta and struct error, the utility of the outside option,
the prices, and the derivative of the utility function wrt prices.
}
\examples{
#NA
}

