% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogitDemandClassZetaFP.R
\name{ldmkt_zeta_fixed_point}
\alias{ldmkt_zeta_fixed_point}
\title{Zeta Fixed Point}
\usage{
ldmkt_zeta_fixed_point(tol)
}
\arguments{
\item{tol}{when should the fixed point converge}
}
\value{
invisibly returns the object (the method updates the prices, market shares, and derivative matrices directly)
}
\description{
Method for the Logit Demand Class that computes the NE prices given marginal cost and information about consumer utility.
This method uses the Zeta Fixed Point of Morrow and Skelos 2010.
}
\examples{
#NA
}

