% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm_profits.R
\name{marginal_cost}
\alias{marginal_cost}
\title{Compute Marginal Costs}
\usage{
marginal_cost(x)
}
\arguments{
\item{x}{a 'Logit_Demand_Market' object}
}
\value{
x with the Market$Mc changed to Market$Price - Market$Markup
}
\description{
Compute Marginal Costs
}
\examples{
#NA
}

