% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm_profits.R
\name{markups}
\alias{markups}
\title{Compute Eq. Markups from BLP Markup equation}
\usage{
markups(x)
}
\arguments{
\item{x}{a 'Logit_Demand_Market' object}
}
\value{
x with equilibrium markups implied by the firms first order condition
}
\description{
Compute Eq. Markups from BLP Markup equation
}
\examples{
#NA
}

