% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm_profits.R
\name{Mc_method}
\alias{Mc_method}
\title{Method for updating Mc}
\usage{
Mc_method(x)
}
\arguments{
\item{x}{a 'logit_demand_market' object}
}
\value{
x with Mc set to Mc_fixed + Mc_error
}
\description{
Method for updating Mc
}
\examples{
#NA
}

