% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OwnershipMatrix.R
\name{O_fun}
\alias{O_fun}
\title{Make Ownership Matrices}
\usage{
O_fun(Firm, Jt)
}
\arguments{
\item{Firm}{logical vector indicating if firm 1 owns product j}

\item{Jt}{the number of products}
}
\value{
a matrix where O[j,k] is one if Firm[j]==Firm[k]
}
\description{
Make Ownership Matrices
}
\examples{
firm <- c(1,1,1,0,0)
jt <- length(Firm)
O <- O_fun(Firm=firm, Jt=jt)
}
