% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassConstructor.R
\name{rldmkt_construct}
\alias{rldmkt_construct}
\title{Random Coef. Logit Market Constructor}
\usage{
rldmkt_construct(Firms, Muij_mat, Aij_mat, Prod_ids = NULL, Delta = NULL,
  Price = NULL, Share = NULL, Struct_error = NULL, Mc_fixed = NULL,
  Mc_error = NULL, Markup = NULL, Alpha = 1, U_out_opt = 0,
  Market_Size = 1)
}
\arguments{
\item{Firms}{vector giving which firm owns each product in the market}

\item{Muij_mat}{number of products x number of simulation draws matrix of random part of consumer utility}

\item{Aij_mat}{number of products x number of simulation draws matrix of random part of consumer's coefficient on price}

\item{Prod_ids}{unique ids for the products in the market}

\item{Delta}{fixed part of consumer utility for each product}

\item{Price}{the prices of each product}

\item{Share}{the market share of each product}

\item{Struct_error}{the econometric error for each product}

\item{Mc_fixed}{the fixed part of marginal cost for each product}

\item{Mc_error}{the econometric error in marginal cost}

\item{Markup}{the markup for each product}

\item{Alpha}{the fixed part of the derivative of the consumers utility function with respect to price}

\item{U_out_opt}{the utility of the outside option (default = 0)}

\item{Market_Size}{the size of the market (default to 1)}
}
\value{
has no return
}
\description{
Construct the random coefficient logit objects.
}
\examples{
#NA
}

