% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassMethods.R
\name{rldmkt_Ds_fun}
\alias{rldmkt_Ds_fun}
\title{Random Ceof. Jacobian of Market Shares}
\usage{
rldmkt_Ds_fun()
}
\value{
invisibly return self
}
\description{
Method for rldmkt class that simulates the jaobian of market shares.
}
\examples{
#NA
}

