% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassGetters.R
\name{rldmkt_getAijs}
\alias{rldmkt_getAijs}
\title{Get Method for Aijs}
\usage{
rldmkt_getAijs()
}
\value{
Aijs 1 x draws vector of random part of coeficient on price
}
\description{
Get Method for Aijs
}
\examples{
#NA
}

