% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassGetters.R
\name{rldmkt_getAlpha}
\alias{rldmkt_getAlpha}
\title{Get Method for Alpha}
\usage{
rldmkt_getAlpha()
}
\value{
fixed part of coefficient on price
}
\description{
Get Method for Alpha
}
\examples{
#NA
}

