% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassGetters.R
\name{rldmkt_getDerivPrice}
\alias{rldmkt_getDerivPrice}
\title{Get method for Coefficient on price}
\usage{
rldmkt_getDerivPrice()
}
\value{
Deriv_price, the sum of the random and fixed parts of the coefficient on price
}
\description{
Get method for Coefficient on price
}
\examples{
#NA
}

