% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassGetters.R
\name{rldmkt_getMuijs}
\alias{rldmkt_getMuijs}
\title{Get method for random part of utility}
\usage{
rldmkt_getMuijs()
}
\value{
Muijs, number of products x draws matrix of the random parts of the random coefficients on product characteristics
}
\description{
Get method for random part of utility
}
\examples{
#NA
}

