% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassGetters.R
\name{rldmkt_getUijs}
\alias{rldmkt_getUijs}
\title{Get Method for Uijs}
\usage{
rldmkt_getUijs()
}
\value{
Uijs number of products x draws matrix of the sume of the fixed part of utility, random coefficients on the product characteristics, and the structural errors
}
\description{
Get Method for Uijs
}
\examples{
#NA
}

