% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLogitDemandClassMethods.R
\name{rldmkt_share}
\alias{rldmkt_share}
\title{Random Coef. Logit Share Function}
\usage{
rldmkt_share()
}
\value{
invisibly self
}
\description{
Method for rldmkt class that simulates market shares.
}
\examples{
#NA
}

