% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{share}
\alias{share}
\title{Compute Market shares}
\usage{
share(x)
}
\arguments{
\item{x}{an object with a share method}
}
\value{
x with the market shares updated
}
\description{
Compute Market shares
}
\examples{
#share(x)
}
\seealso{
\code{\link{share.Logit_Demand_Market}}
}
