% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{zeta_fixed_point}
\alias{zeta_fixed_point}
\title{Zeta Fixed Point Iterations}
\usage{
zeta_fixed_point(x, tol)
}
\arguments{
\item{x}{object with a zeta fixed point method}
}
\value{
x updated with prices using the zeta fixed point iterations
}
\description{
Zeta Fixed Point Iterations
}
\examples{
#zeta_fixed_point(x)
}
\seealso{
\code{\link{zeta_fixed_point.Logit_Demand_Market}}
}
